# admin_openapi.yaml
openapi: 3.0.3
info:
  title: Admin BFF
  version: 1.0.0
  description: |
    API gateway for players, wallets, withdrawals, and reports services.
    Conventions:
      - Array fields are always arrays (empty array when no data).
      - Optional fields may be null when data is unavailable.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /users/api/v1/staffs/logout:
    delete:
      tags: [Auth]
      summary: Logout (stateless)
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/api/v1/staffs:
    get:
      tags: [Auth]
      summary: List users
      parameters:
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
              examples:
                empty:
                  summary: Empty list
                  value:
                    items: []
                    total: 0
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Auth]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateStaffUserResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /users/api/v1/staffs/{id}/ban:
    put:
      tags: [Auth]
      summary: Block user
      parameters:
        - $ref: "#/components/parameters/idInt"
      responses:
        "200":
          description: Blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/api/v1/staffs/me:
    get:
      tags: [Auth]
      summary: Get current user
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
              examples:
                withEmptyRoles:
                  summary: No roles assigned
                  value:
                    data:
                      username: "admin"
                      roles: []
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    idInt:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    offset:
      name: offset
      in: query
      schema:
        type: integer
    limit:
      name: limit
      in: query
      schema:
        type: integer
...
  schemas:
...
    UserProfile:
      type: object
      properties:
        data:
          type: object
          properties:
            username:
              type: string
            roles:
              type: array
              items:
                type: string
              minItems: 0
              default: []
    UsersListResponse:
      type: object
      properties:
        items:
          type: array
          minItems: 0
          default: []
          items:
            type: object
            properties:
              id:
                type: integer
              username:
                type: string
              roles:
                type: array
                items:
                  type: string
                minItems: 0
                default: []
              isBlocked:
                type: boolean
        total:
          type: integer
    CreateStaffUserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        roles:
          type: array
          items:
            type: string
          minItems: 0
          default: []
      required: [username, password]
    CreateStaffUserResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
            roles:
              type: array
              items:
                type: string
              minItems: 0
              default: []
            isBlocked:
              type: boolean
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string